Organising imports

+++ src/main/java/com/facilio/bmsconsole/actions/ContractsAction.java
@@ -11,7 +11,6 @@ import com.facilio.bmsconsole.commands.TransactionChainFactory;
 import com.facilio.bmsconsole.context.ContractAssociatedAssetsContext;
 import com.facilio.bmsconsole.context.ContractAssociatedTermsContext;
 import com.facilio.bmsconsole.context.ContractsContext;
-import com.facilio.bmsconsole.context.LabourContractContext;
 import com.facilio.chain.FacilioContext;
 import com.facilio.constants.FacilioConstants;
 
@@ -122,7 +121,7 @@ public class ContractsAction extends FacilioAction{
 		chain.execute(context);
 		
 		if (getFetchCount()) {
-			setResult(FacilioConstants.ContextNames.RECORD_COUNT,(Long) context.get(FacilioConstants.ContextNames.RECORD_COUNT));
+			setResult(FacilioConstants.ContextNames.RECORD_COUNT,context.get(FacilioConstants.ContextNames.RECORD_COUNT));
 		}
 		else {
 			List<ContractsContext> contracts = (List<ContractsContext>) context.get(FacilioConstants.ContextNames.RECORD_LIST);
