+++ src/main/java/com/facilio/bmsconsole/actions/LabourContractAction.java
@@ -114,7 +114,7 @@ public class LabourContractAction extends FacilioAction{
 		chain.execute(context);
 		
 		if (getFetchCount()) {
-			setResult(FacilioConstants.ContextNames.RECORD_COUNT,(Long) context.get(FacilioConstants.ContextNames.RECORD_COUNT));
+			setResult(FacilioConstants.ContextNames.RECORD_COUNT,context.get(FacilioConstants.ContextNames.RECORD_COUNT));
 		}
 		else {
 			List<LabourContractContext> labourContracts = (List<LabourContractContext>) context.get(FacilioConstants.ContextNames.RECORD_LIST);
