+++ src/main/java/com/facilio/bmsconsole/actions/RentalLeaseContractAction.java
@@ -160,7 +160,7 @@ public class RentalLeaseContractAction extends FacilioAction{
 		chain.execute(context);
 		
 		if (getFetchCount()) {
-			setResult(FacilioConstants.ContextNames.RECORD_COUNT,(Long) context.get(FacilioConstants.ContextNames.RECORD_COUNT));
+			setResult(FacilioConstants.ContextNames.RECORD_COUNT,context.get(FacilioConstants.ContextNames.RECORD_COUNT));
 		}
 		else {
 			List<RentalLeaseContractContext> rentalLeaseContracts = (List<RentalLeaseContractContext>) context.get(FacilioConstants.ContextNames.RECORD_LIST);
