+++ src/main/java/com/facilio/bmsconsole/actions/ServiceAction.java
@@ -130,7 +130,7 @@ public class ServiceAction extends FacilioAction{
 		chain.execute(context);
 		
 		if (getFetchCount()) {
-			setResult(FacilioConstants.ContextNames.RECORD_COUNT,(Long) context.get(FacilioConstants.ContextNames.RECORD_COUNT));
+			setResult(FacilioConstants.ContextNames.RECORD_COUNT,context.get(FacilioConstants.ContextNames.RECORD_COUNT));
 		}
 		else {
 			List<ServiceContext> serviceContexts = (List<ServiceContext>) context.get(FacilioConstants.ContextNames.RECORD_LIST);
